langsmith/evaluation/requirements.txt
langsmith==0.1.142
openai==1.54.4
packaging==24.2
python-dotenv==1.0.1